split trait linting into its own class
